Fixed the return value of play_sfall_sound function
* before it returned a raw pointer instead of the ID number of the
played sound.

+++ sfall/Modules/Scripting/Handlers/Misc.cpp
@@ -616,7 +616,7 @@ void sf_get_attack_type(OpcodeContext& ctx) {
 }
 
 void sf_play_sfall_sound(OpcodeContext& ctx) {
-	long soundID = (long)Sound::PlaySfallSound(ctx.arg(0).strValue(), ctx.arg(1).asBool());
+	DWORD soundID = Sound::PlaySfallSound(ctx.arg(0).strValue(), ctx.arg(1).asBool());
 	ctx.setReturn(soundID);
 }
